refactor: user EnvironmentScope only as request scope
